# Optimize vector::assign for InputIterator-only pair inputs