--- a/libcxx/docs/ReleaseNotes/20.rst
+++ b/libcxx/docs/ReleaseNotes/20.rst
@@ -69,9 +69,8 @@ Improvements and New Features
 - The ``_LIBCPP_ABI_BOUNDED_ITERATORS_IN_STD_ARRAY`` ABI configuration was added, which allows storing valid bounds
   in ``std::array::iterator`` and detecting OOB accesses when the appropriate hardening mode is enabled.
 
-- The `assign(_InputIterator, _InputIterator)` function of `std::vector<_Tp, _Allocator>` has been optimized for
-  non-trivial element types, such as `std::vector<std::string>`, with a performance improvement of up to 2.3x. The
-  performance for trivial types, such as `std::vector<int>`, remains similar or shows slight improvements.
+- The input iterator overload of `assign(_InputIterator, _InputIterator)` in `std::vector<_Tp, _Allocator>` has been 
+  optimized, resulting in a performance improvement of up to 3.7x.
 
 Deprecations and Removals
 -------------------------