--- a/libcxx/test/benchmarks/ContainerBenchmarks.h
+++ b/libcxx/test/benchmarks/ContainerBenchmarks.h
@@ -51,27 +51,11 @@ void BM_Assignment(benchmark::State& st, Container) {
   }
 }
 
-template <class Container,
-          class GenInputs,
-          typename std::enable_if<std::is_trivial<typename Container::value_type>::value>::type* = nullptr>
+template <std::size_t... sz, typename Container, typename GenInputs>
 void BM_AssignInputIterIter(benchmark::State& st, Container c, GenInputs gen) {
-  auto in = gen(st.range(1));
-  c.resize(st.range(0));
-  benchmark::DoNotOptimize(&in);
-  benchmark::DoNotOptimize(&c);
-  for (auto _ : st) {
-    c.assign(cpp17_input_iterator(in.begin()), cpp17_input_iterator(in.end()));
-    benchmark::ClobberMemory();
-  }
-}
-
-template <class Container,
-          class GenInputs,
-          typename std::enable_if<!std::is_trivial<typename Container::value_type>::value>::type* = nullptr>
-void BM_AssignInputIterIter(benchmark::State& st, Container c, GenInputs gen) {
-  auto v = gen(1, 100);
+  auto v = gen(1, sz...);
   c.resize(st.range(0), v[0]);
-  auto in = gen(st.range(1), 32);
+  auto in = gen(st.range(1), sz...);
   benchmark::DoNotOptimize(&in);
   benchmark::DoNotOptimize(&c);
   for (auto _ : st) {